Gallery: Automatically get alt text from photo "Caption" field.
Fixes: Consider extracting alt text from "caption" property of image
Fixes #30

"Caption" is the most suitable for an alt-text like description. When this field is set in the photo metadata it is automatically used as alt text. This can be overridden by setting `alt` in photo-metadata. If no caption no custom alt text is set we fall back to title.